add cats typeclass instances
